[AEROGEAR-2732] Add alert widget to mobile client overview
Show an alert widget when there are mobile services available to be
bound. Add a link to the mobile services tab in the mobile client view
within the alert message. The widget should be dismissable and displayed
again when the page reloads.

Also includes a minor fix to use the serviceName instead of the
displayName for the generatedName in the annotations as this fails for
the custom runtime connector.